backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend service for xgd-excel-useful"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
